requirements.txt
# Dependencies for the customer service application
pydantic-settings>=2.8.1
tabulate>=0.9.0
cloudpickle>=3.1.1
google-cloud-aiplatform[adk,agent_engine]>=1.88.0
streamlit>=1.28.0
SQLAlchemy>=2.0.0
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0